chore: Edit Member
- Set Id Long type

+++ src/main/kotlin/me/bossm0n5t3r/blog/member/domain/Member.kt
@@ -17,7 +17,7 @@ class Member(
 
     @Column(name = "name", nullable = false, updatable = true)
     var name: String
-) : BaseEntity<Int>() {
+) : BaseEntity<Long>() {
     @Column(name = "password", nullable = false, updatable = true)
     var password: String = encryptPassword(email, password, name)
